Move name alias to ids
