--- src/main/kotlin/com/github/derg/transpiler/core/Typedefs.kt
@@ -1,3 +1,0 @@
-package com.github.derg.transpiler.core
-
-typealias Name = String
