+++ src/main/kotlin/com/github/derg/transpiler/source/Ids.kt
@@ -1,6 +1,11 @@
 package com.github.derg.transpiler.source
 
 /**
+ * Many objects are given a name which identifies them in some meaningful manner.
+ */
+typealias Name = String
+
+/**
  * All symbols must be granted a unique id to be able to resolve any identifier to an appropriate object. This id will
  * uniquely identify not only the specific object, but also the specific type of object.
  */
