+++ src/main/kotlin/com/github/derg/transpiler/source/ast/Definitions.kt
@@ -1,7 +1,7 @@
 package com.github.derg.transpiler.source.ast
 
-import com.github.derg.transpiler.core.Name
 import com.github.derg.transpiler.source.Mutability
+import com.github.derg.transpiler.source.Name
 import com.github.derg.transpiler.source.Visibility
 
 /**
