+++ src/test/kotlin/com/github/derg/transpiler/phases/parser/Helpers.kt
@@ -1,8 +1,8 @@
 package com.github.derg.transpiler.phases.parser
 
-import com.github.derg.transpiler.core.Name
 import com.github.derg.transpiler.phases.lexer.tokenize
 import com.github.derg.transpiler.source.Mutability
+import com.github.derg.transpiler.source.Name
 import com.github.derg.transpiler.source.Visibility
 import com.github.derg.transpiler.source.ast.*
 import com.github.derg.transpiler.source.ast.Function
